backend/requirements.txt
Flask
flask-cors
yfinance
beautifulsoup4
requests
python-dotenv
Bharat-sm-data